Move pipeline unit tests to subfolder
