+++ tests/init.js
@@ -66,9 +66,9 @@ export const MAX_MODEL_DISPOSE_TIME = 1_000; // 1 second
 
 export const MAX_TEST_TIME = MAX_MODEL_LOAD_TIME + MAX_TEST_EXECUTION_TIME + MAX_MODEL_DISPOSE_TIME;
 
-export const DEFAULT_MODEL_OPTIONS = {
+export const DEFAULT_MODEL_OPTIONS = Object.freeze({
   dtype: "fp32",
-};
+});
 
 expect.extend({
   toBeCloseToNested(received, expected, numDigits = 2) {
